Moving request modifiers to its own file
